Update index.html
The necessary work in `vite.config.js` that this is referring to appears to have already been done! However, it's quite valuable to keep an index.html here in case someone goes to the wrong localhost port in their dev server, so this doesn't delete the file, it just changes its purpose (and removes vite imports, which IMO avoids accidental confusion).